update
added experimental functionality to the printing module, ability to write streams from individual MPI ranks to their own files. Useful for parameter scans where progress is not expected to be similar.

Added custom domain decomposition option

Added more to the main block. May be simpler to merge the `extra` module at this point...

Exposed the temperature options when creating a gridloaded particle bunch, enables an energy spread and emittance to be included.